// src/types/common.ts
export interface IFormInput {
  serviceName:string
  hotelName: string;
  cityId: string;
  hotelDescription: string;
  contact: string;
  price: number;
  discount: number;
  latitude: number;
  longitude: number;
  oldPassword: string;
  newPassword: string;
  categoryName:string;
  description:string
  overview:string
  confirm: string;
  email: string;
  roomSize: string;
  roomName: string;
  bedDetail: string;
  subCatName:string
  adults: number;
}